+++ test/resampling/resampling_schemes/test_BinnedResampling.jl
@@ -3,3 +3,4 @@ grid = 0:10:500
 resampling = BinnedResampling(grid, n_draws)
 
 @test resampling isa BinnedResampling
+@test typeof(resampling) <: AbstractBinnedUncertainValueResampling